refactor: change path script console color
