+++ src/fs/add.js
@@ -1,6 +1,6 @@
 import fs from "fs";
-import { log } from "../utils/colorConsole/colorConsole.js";
 import path from "path";
+import { log } from "../utils/colorConsole.js";
 
 export const add = (newFileName, currentDir) => {
   log.cyan("run add");
